+++ core/api/golang/lib/enclaves/enclave_context.go
@@ -531,6 +531,17 @@ func (enclaveCtx *EnclaveContext) StoreWebFiles(ctx context.Context, urlToStoreW
 	return services.FilesArtifactID(response.Uuid), nil
 }
 
+// Docs available at https://docs.kurtosistech.com/kurtosis-core/lib-documentation
+func (EnclaveContext *EnclaveContext) StoreFilesFromService(ctx context.Context, serviceId services.ServiceID, absoluteFilepathOnServiceContainer string) (services.FilesArtifactID, error) {
+	serviceIdStr := string(serviceId)
+	args := binding_constructors.NewStoreFilesArtifactFromServiceArgs(serviceIdStr, absoluteFilepathOnServiceContainer)
+	response, err := EnclaveContext.client.StoreFilesArtifactFromService(ctx, args)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred copying source content from absolute filepath '%v' in service container with ID '%v'", absoluteFilepathOnServiceContainer, serviceIdStr)
+	}
+	return services.FilesArtifactID(response.Uuid), nil
+}
+
 // ====================================================================================================
 // 									   Private helper methods
 // ====================================================================================================